+++ sample-apps/react/livestream-app/src/viewers/ui/ViewerHeader.tsx
@@ -5,7 +5,7 @@ import {
   HDBadge,
   LiveBadge,
   TotalViewersBadge,
-} from '../../ui/components';
+} from '../../components';
 import {
   IconButton,
   MenuToggle,
